requirements.txt
streamlit
openai
faiss-cpu
numpy
python-dotenv
requests
sentence-transformers
biopython
